+++ tests/test_public_api.py
@@ -5,9 +5,12 @@ import pytest
 import sunnbear._core.exceptions
 import sunnbear._core.functions.core
 import sunnbear._core.functions.core.exceptions
+import sunnbear._core.solvers.core
+import sunnbear._core.solvers.core.exceptions
 import sunnbear._core.stats
 import sunnbear.exceptions
 import sunnbear.functions
+import sunnbear.solvers
 import sunnbear.stats
 
 
@@ -30,11 +33,13 @@ def _is_submodule(module, name: str) -> bool:
     [
         (sunnbear.stats, (sunnbear._core.stats,)),
         (sunnbear.functions, (sunnbear._core.functions.core,)),
+        (sunnbear.solvers, (sunnbear._core.solvers.core,)),
         (
             sunnbear.exceptions,
             (
                 sunnbear._core.exceptions,
                 sunnbear._core.functions.core.exceptions,
+                sunnbear._core.solvers.core.exceptions,
             ),
         ),
     ],